Git commit to change the structure of header files in Arduino-Libraries.
As arduino-cli compiles scripts by finding them recursively, the relative paths used caused the compiler to fail.
However the relative paths are needed by gcc, to avoid using -I includes in all makefiles.

So to avoid this, a preprocessor instruction has been added that will insert the relative paths if Arduino_h is not defined.

This way boths compilers will work and the code can be used both on a computer and in an arduino.